fixed bug, readings per day did not display correct.

+++ ggc/src/datamodels/HbA1cValues.java
@@ -93,7 +93,7 @@ public class HbA1cValues
     public float getReadingsPerDay()
     {
         if (dayCount != 0)
-            return readings / dayCount;
+            return readings / (float)dayCount;
         else
             return 0;
     }
